// src/utils/FetchBackend/rest/api/users/dto/create-user.dto.ts
export default interface CreateUserDto {
  dp_login: string;
  dp_email: string;
  dp_password: string;
  dp_unp: string;
  dp_nameLegalEntity: string;
  dp_shortNameLegalEntity: string;
  dp_address: string;
  dp_receptionPhone: string;
  dp_firstName: string;
  dp_lastName: string;
  dp_middleName: string;
}